\name{AdequacyModel-package}
\alias{AdequacyModel-package}
\encoding{latin1}
\alias{AdequacyModel}
\docType{package}

\title{
Adequacy of probabilistic models and generation of pseudo-random numbers.
}

\description{This package provides some useful functions for calculating quality measures for adjustment, including statistics Cramer-von Mises and Anderson-Darling. These statistics are often used to compare models not equipped. Also provided are other measures of fitness, such as AIC, CAIC and BIC. The package also provides functions for generating pseudo-random number of random variables that follow the following probability distributions: exponentiated Weibull, Kumaraswamy Weibull, Kumaraswamy Weibull Poisson.
}
\details{
\tabular{ll}{
Package: \tab AdequacyModel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-15\cr
License: \tab GPL-2\cr
Depends: \tab R (>= 2.10.0)\cr
}
}

\references{
Chen, G., Balakrishnan, N. (1995). A general purpose approximate goodness-of-fit test. Journal of Quality Technology 27, 154-161.
}
