\name{carbone}
\alias{carbone}
\encoding{latin1}
\title{
Breaking stress of carbon fibres
}
\description{
The first real data set corresponds to an uncensored data set from Nichols and Padgett (2006) on breaking stress of carbon fibres (in Gba).
}
\usage{data(carbone)}
\format{
  The format is:
 num [1:100] 3.7 2.74 2.73 2.5 3.6 3.11 3.27 2.87 1.47 3.11 ...
}

\references{
Nichols, M.D, Padgett, W.J. (2006). A Bootstrap control chart for Weibull percentiles. Quality and Reliability Engineering International 22, 141-151.
}
\examples{
data(carbone)
hist(carbone)
}
\keyword{datasets}
