\name{AdequacyModel-package}
\alias{AdequacyModel-package}
\encoding{UTF-8}
\alias{AdequacyModel}
\docType{package}

\title{
Adequacy of probabilistic models and and general purpose optimization
}

\description{The main application concerns to a new robust optimization package with two major contributions. The first contribution refers to the assessment of the adequacy of probabilistic models through a combination of several statistics, which measure the relative quality of statistical models for a given data set. The second one provides a general purpose optimization method based on meta-heuristics functions for maximizing or minimizing an arbitrary objective function.}
\details{
\tabular{ll}{
Package: \tab AdequacyModel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
%Date: \tab 2012-11-15\cr
License: \tab GPL-2\cr
Depends: \tab R (>= 2.10.0)\cr
}
}

\references{
Aarset, M. V. (1987). How to identify bathtub hazard rate. IEEE Transactions Reliability, 36, 106-108.

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Beni G, Wang J (1993). Swarm intelligence in cellular robotic systems. pp. 703-712.

Chen, G., Balakrishnan, N. (1995). A general purpose approximate goodness-of-fit test. Journal of Quality Technology 27, 154-16.

Eberhart RC, Kennedy J (1995). A new optimizer using particle swarm theory. In Proceedings of the sixth international symposium on micro machine and human science, volume 1, pp. 39-43. New York, NY.

Kennedy J, Kennedy JF, Eberhart RC, Shi Y (2001). Swarm intelligence. Morgan Kaufmann.

Nichols, M.D, Padgett, W.J. (2006). A Bootstrap control chart for Weibull percentiles. Quality and Reliability Engineering International 22, 141-151.

Shi Y, Eberhart R (1998). A modified particle swarm optimizer. In Evolutionary Computation Proceedings, 1998. IEEE World Congress on Computational Intelligence., The 1998 IEEE International Conference on, pp. 69-73. IEEE.
}
