\name{AdequacyModel-package}
\alias{AdequacyModel-package}
\encoding{latin1}
\alias{AdequacyModel}
\docType{package}

\title{
Adequacy of models
}

\description{
This package provides some useful functions for calculating measures of goodness of fit, including statistics Cramer-von Mises and Anderson-Darling. These statistics are often  used to compare models not fitted. You can also calculate other medias goodness of fit  as the AIC, CAIC and BIC.
}
\details{
\tabular{ll}{
Package: \tab AdequacyModel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-15\cr
License: \tab GPL-2\cr
}
}

\references{
Chen, G., Balakrishnan, N. (1995). A general purpose approximate goodness-of-fit test. Journal of Quality Technology 27, 154-161.
}
